authentication and redirect works, if there is no token you will be redirected to login

+++ frontend/src/components/Post/Post.jsx
@@ -1,11 +1,20 @@
 import { usePostsContext } from '../../hooks/usePostsContext';
+import { useAuthContext } from '../../hooks/useAuthContext';
 
 const Post = ({ post }) => {
     const { dispatch } = usePostsContext();
+    const { user } = useAuthContext();
 
     const handleDelete = async() => {
+        if (!user) {
+            return;
+        }
+
         const response = await fetch('/api/protected/post/' + post._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
         const json = await response.json();
